clam: daily update and scan
The whole home directory is scanned once a day using crontab. Copy this
file into /etc/cron.daily. If no threts are found you get notification
about that. If some are found they are logged into a file. The
notification gives you an option to go to the directory and look at the
log. If it was a mistake let it be otherwise you can delete the files.

If you want the clamscan to delete the infected files add '--remove'
option to the scan.

Signed-off-by: Filip Lobpreis <f.lobpreis@gmail.com>